@@ -1,30 +1,41 @@
 {
-  "name": "anistand-frontend",
-  "version": "1.0.0",
+  "name": "anistand",
   "private": true,
+  "version": "1.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "test:connection": "node scripts/test-connection.js"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
   },
   "dependencies": {
-    "@types/react": "^18.2.37",
-    "@types/react-dom": "^18.2.15",
-    "autoprefixer": "^10.4.16",
-    "axios": "^1.12.2",
+    "@reduxjs/toolkit": "^2.0.1",
+    "axios": "^1.6.2",
     "lucide-react": "^0.294.0",
-    "next": "^14.2.33",
-    "postcss": "^8.4.32",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tailwindcss": "^3.3.6",
-    "typescript": "^5.2.2"
+    "react-redux": "^9.0.4",
+    "react-router-dom": "^6.20.1",
+    "video.js": "^8.6.1",
+    "videojs-contrib-quality-levels": "^3.0.0",
+    "videojs-hls-quality-selector": "^1.1.4"
   },
   "devDependencies": {
-    "@types/node": "^24.6.2",
-    "eslint": "^8.54.0",
-    "eslint-config-next": "14.0.3"
+    "@types/node": "^20.10.5",
+    "@types/react": "^18.2.43",
+    "@types/react-dom": "^18.2.17",
+    "@types/video.js": "^7.3.56",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.55.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.3.6",
+    "typescript": "^5.2.2",
+    "vite": "^7.1.9"
   }
 }
